Add edition tools state machine
Add edition action history using command pattern
Fix tiles selection bug : while editing one tile height, once it position is too far, automating select another tile even when the mouse hasn't move
Fix: Block view zoom or rotation while editing ; allow editing while view is translating

+++ src/model/TileCorner.hpp
@@ -16,12 +16,12 @@ public:
     float getHeight() const;
     int getTextureID() const;
     sf::Vector2f getPosition() const;
-    void addHeight(int heightStep);
+    void addHeight(float heightStep);
     void setTextureId(int textureId);
 private:    
     int m_row;
     int m_col;
-    int m_height;
+    float m_height;
     int m_textureID;
 };
